refactor: refatorando os cards,  adapters e as classes fragments em cada tela e criando pastas para organização
