--- app/src/main/java/com/example/scaneia/ModelosFragment.java
@@ -1,48 +1,0 @@
-package com.example.scaneia;
-
-import android.content.Intent;
-import android.os.Bundle;
-import androidx.annotation.NonNull;
-import androidx.fragment.app.Fragment;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.ImageView;
-import com.example.scaneia.databinding.FragmentModelosBinding;
-
-public class ModelosFragment extends Fragment {
-
-    private FragmentModelosBinding binding;
-
-    public View onCreateView(@NonNull LayoutInflater inflater,
-                             ViewGroup container, Bundle savedInstanceState) {
-
-        binding = FragmentModelosBinding.inflate(inflater, container, false);
-        View root = binding.getRoot();
-
-        RecyclerView recyclerView = binding.rvModelos;
-        AdapterModelos adapter = new AdapterModelos(10);
-
-        recyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
-        recyclerView.setAdapter(adapter);
-
-        ImageView profile = root.findViewById(R.id.profile);
-        profile.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(getActivity(), Profile.class);
-                startActivity(intent);
-            }
-        });
-
-        return root;
-    }
-
-    @Override
-    public void onDestroyView() {
-        super.onDestroyView();
-        binding = null;
-    }
-}
