+++ app/src/main/java/com/example/scaneia/analises/AnalisesFragment.java
@@ -1,4 +1,4 @@
-package com.example.scaneia.ui.home;
+package com.example.scaneia.analises;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -6,41 +6,29 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
-
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
-
-import com.example.scaneia.AdapterPlanilha;
-import com.example.scaneia.Profile;
+import com.example.scaneia.Perfil;
 import com.example.scaneia.R;
-import com.example.scaneia.databinding.FragmentHomeBinding;
+import com.example.scaneia.databinding.FragmentAnalisesBinding;
 
-public class HomeFragment extends Fragment {
+public class AnalisesFragment extends Fragment {
 
-    private FragmentHomeBinding binding;
+    private FragmentAnalisesBinding binding;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        binding = FragmentHomeBinding.inflate(inflater, container, false);
+        binding = FragmentAnalisesBinding.inflate(inflater, container, false);
         View root = binding.getRoot();
-
-        RecyclerView recyclerView = binding.rv;
-        AdapterPlanilha adapter = new AdapterPlanilha(5);
-
-        recyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
-        recyclerView.setAdapter(adapter);
 
         ImageView profile = root.findViewById(R.id.profile);
         profile.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Intent intent = new Intent(getActivity(), Profile.class);
+                Intent intent = new Intent(getActivity(), Perfil.class);
                 startActivity(intent);
             }
         });
-
         return root;
     }
