+++ app/src/main/java/com/example/scaneia/funcionarios/FuncionariosFragment.java
@@ -1,4 +1,4 @@
-package com.example.scaneia.ui.notifications;
+package com.example.scaneia.funcionarios;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -6,34 +6,42 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
-import android.widget.TextView;
-
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
-import androidx.lifecycle.ViewModelProvider;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
-import com.example.scaneia.Profile;
+import com.example.scaneia.Perfil;
 import com.example.scaneia.R;
-import com.example.scaneia.databinding.FragmentNotificationsBinding;
+import com.example.scaneia.databinding.FragmentFuncionariosBinding;
 
-public class NotificationsFragment extends Fragment {
+public class FuncionariosFragment extends Fragment {
 
-    private FragmentNotificationsBinding binding;
+    private FragmentFuncionariosBinding binding;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        binding = FragmentNotificationsBinding.inflate(inflater, container, false);
+        binding = FragmentFuncionariosBinding.inflate(inflater, container, false);
         View root = binding.getRoot();
+
+
+        RecyclerView recyclerView = binding.rv;
+        AdapterFuncionarios adapter = new AdapterFuncionarios(10);
+
+        recyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
+        recyclerView.setAdapter(adapter);
 
         ImageView profile = root.findViewById(R.id.profile);
         profile.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Intent intent = new Intent(getActivity(), Profile.class);
+                Intent intent = new Intent(getActivity(), Perfil.class);
                 startActivity(intent);
             }
         });
+
         return root;
+
     }
 
     @Override
