+++ app/src/main/java/com/example/scaneia/modelos/ModelosFragment.java
@@ -1,37 +1,32 @@
-package com.example.scaneia.ui.dashboard;
+package com.example.scaneia.modelos;
 
 import android.content.Intent;
 import android.os.Bundle;
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
-import android.widget.TextView;
 
-import androidx.annotation.NonNull;
-import androidx.fragment.app.Fragment;
-import androidx.lifecycle.ViewModelProvider;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
-
-import com.example.scaneia.AdapterFuncionarios;
-import com.example.scaneia.AdapterPlanilha;
-import com.example.scaneia.Profile;
+import com.example.scaneia.Perfil;
 import com.example.scaneia.R;
-import com.example.scaneia.databinding.FragmentDashboardBinding;
+import com.example.scaneia.databinding.FragmentModelosBinding;
 
-public class DashboardFragment extends Fragment {
+public class ModelosFragment extends Fragment {
 
-    private FragmentDashboardBinding binding;
+    private FragmentModelosBinding binding;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        binding = FragmentDashboardBinding.inflate(inflater, container, false);
+
+        binding = FragmentModelosBinding.inflate(inflater, container, false);
         View root = binding.getRoot();
 
-
-        RecyclerView recyclerView = binding.rv;
-        AdapterFuncionarios adapter = new AdapterFuncionarios(5);
+        RecyclerView recyclerView = binding.rvModelos;
+        AdapterModelos adapter = new AdapterModelos(10);
 
         recyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
         recyclerView.setAdapter(adapter);
@@ -40,13 +35,12 @@ public class DashboardFragment extends Fragment {
         profile.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Intent intent = new Intent(getActivity(), Profile.class);
+                Intent intent = new Intent(getActivity(), Perfil.class);
                 startActivity(intent);
             }
         });
 
         return root;
-
     }
 
     @Override
